# schemas/openapi-v1alpha1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agent-computer control API",
    "version": "0.1.0",
    "description": "Development control API: static validation, revocable service credentials, independent runtime grants, authorized definition plans, atomic apply and durable coordination progress. Computer runtime admission, OIDC and execution remain unsupported."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1alpha1/capabilities": {
      "get": {
        "operationId": "capabilities",
        "security": [],
        "responses": {
          "200": {
            "description": "Implemented and unsupported capabilities.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1alpha1/openapi.json": {
      "get": {
        "operationId": "openapi",
        "security": [],
        "responses": {
          "200": {
            "description": "This OpenAPI document.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1alpha1/definitions/validate": {
      "post": {
        "operationId": "validateDefinition",
        "description": "Authenticated, side-effect-free static JSON validation. Requires definitions.validate. No resources are applied. Body limit 1 MiB. Browser Origin and content encoding are rejected. This query does not require an Idempotency-Key.",
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputerSet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Static validation report; this does not grant authority.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "valid",
                    "definition_digest",
                    "diagnostics",
                    "scope"
                  ],
                  "properties": {
                    "valid": {
                      "const": true
                    },
                    "definition_digest": {
                      "type": "string",
                      "pattern": "^sha256:[0-9a-f]{64}$"
                    },
                    "diagnostics": {
                      "type": "array"
                    },
                    "scope": {
                      "const": "static"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unreadable body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired, disabled or revoked credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope missing or browser authentication unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Request timeout.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 1 MiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported content type or encoding.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid ComputerSet. details.validation contains diagnostics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable or capacity limit reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1alpha1/plans": {
      "post": {
        "operationId": "createDefinitionPlan",
        "description": "Create/retry an immutable diff and impact plan. Writes only the plan, receipt and event/Outbox; no resource definitions or runtime effects. Existing root metadata and each existing resource require expectedRevision. Requires definitions.manage and current definition/reference grants. Browser Origin is rejected.",
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "responses": {
          "201": {
            "description": "Create/retry an immutable diff and impact plan. Writes only the plan, receipt and event/Outbox; no resource definitions or runtime effects. Existing root metadata and each existing resource require expectedRevision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionPlan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, missing/invalid Idempotency-Key, or malformed apply body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired, disabled or revoked credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing definitions.manage, object grant or unsupported browser Origin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Request timeout. The commit outcome may be unknown; retry with the same key and input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency input conflict or plan digest mismatch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Retired request key or expired unapplied plan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "412": {
            "description": "Declaration/resource revision or referenced head changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 1 MiB or resolved plan exceeds 8 MiB / dependency traversal bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported content type or content encoding.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid declaration, unavailable reference (missing, wrong kind, disabled or unauthorized), or unsupported data migration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "428": {
            "description": "An existing declaration or resource is missing expectedRevision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable or capacity limit reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputerSet"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ]
      }
    },
    "/v1alpha1/plans/{id}": {
      "get": {
        "operationId": "getDefinitionPlan",
        "description": "Read the caller-owned plan under current original plan permissions, including transitive reference grants. Requires definitions.manage and current definition/reference grants. Browser Origin is rejected.",
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Read the caller-owned plan under current original plan permissions, including transitive reference grants.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionPlan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, missing/invalid Idempotency-Key, or malformed apply body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired, disabled or revoked credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing definitions.manage, object grant or unsupported browser Origin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Plan/operation not found in the authenticated principal and organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Request timeout. The commit outcome may be unknown; retry with the same key and input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "412": {
            "description": "Declaration/resource revision or referenced head changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 1 MiB or resolved plan exceeds 8 MiB / dependency traversal bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid declaration, unavailable reference (missing, wrong kind, disabled or unauthorized), or unsupported data migration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable or capacity limit reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^plan_[0-9a-f]{32}$"
            }
          }
        ]
      }
    },
    "/v1alpha1/plans/{id}/apply": {
      "post": {
        "operationId": "applyDefinitionPlan",
        "description": "Atomically publish immutable versions, creator grants, Queued operation, Pending intents, receipt and event/Outbox. Rechecks authorization, first-apply expiry and all selected revisions. Repeating an applied plan returns the same operation with current progress, including under a new key. Requires definitions.manage and current definition/reference grants. Browser Origin is rejected.",
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "responses": {
          "202": {
            "description": "Atomically publish immutable versions, creator grants, Queued operation, Pending intents, receipt and event/Outbox. Rechecks authorization, first-apply expiry and all selected revisions. Repeating an applied plan returns the same operation with current progress, including under a new key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionOperation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, missing/invalid Idempotency-Key, or malformed apply body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired, disabled or revoked credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing definitions.manage, object grant or unsupported browser Origin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Plan/operation not found in the authenticated principal and organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Request timeout. The commit outcome may be unknown; retry with the same key and input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency input conflict or plan digest mismatch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Retired request key or expired unapplied plan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "412": {
            "description": "Declaration/resource revision or referenced head changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 1 MiB or resolved plan exceeds 8 MiB / dependency traversal bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported content type or content encoding.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid declaration, unavailable reference (missing, wrong kind, disabled or unauthorized), or unsupported data migration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable or capacity limit reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyDefinitionPlan"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^plan_[0-9a-f]{32}$"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ]
      }
    },
    "/v1alpha1/operations/{id}": {
      "get": {
        "operationId": "getDefinitionOperation",
        "description": "Read the caller-owned operation and its current state under current original plan permissions. Requires definitions.manage and current definition/reference grants. Browser Origin is rejected.",
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Read the caller-owned operation and its current state under current original plan permissions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionOperation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, missing/invalid Idempotency-Key, or malformed apply body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired, disabled or revoked credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing definitions.manage, object grant or unsupported browser Origin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Plan/operation not found in the authenticated principal and organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Request timeout. The commit outcome may be unknown; retry with the same key and input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "412": {
            "description": "Declaration/resource revision or referenced head changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 1 MiB or resolved plan exceeds 8 MiB / dependency traversal bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid declaration, unavailable reference (missing, wrong kind, disabled or unauthorized), or unsupported data migration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable or capacity limit reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^op_[0-9a-f]{32}$"
            }
          }
        ]
      }
    },
    "/v1alpha1/runtime-access/{kind}/{id}": {
      "get": {
        "operationId": "getRuntimeAccess",
        "description": "Inspect the authenticated principal own effective permissions. Requires runtime.read and an exact read grant on this target. Organization and principal are derived solely from the service credential. Browser Origin is rejected. The response grants no future execution authority.",
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "computer",
                "workspace",
                "app",
                "browser_profile"
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current effective permissions for the authenticated principal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeAccess"
                }
              }
            }
          },
          "400": {
            "description": "Invalid target kind or resource ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired, revoked or disabled credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing runtime.read scope or unsupported browser Origin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource is missing, disabled, in another organization, or lacks the caller read grant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Only GET is accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Request deadline exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "serviceBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "acsk_<credential-id>_<secret>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "request_id",
          "details"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "request_id": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
...
      "ApplyDefinitionPlan": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "plan_digest"
        ],
        "properties": {
          "plan_digest": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$"
          }
        }
      },
...
    },
    "responses": {
      "Unavailable": {
        "description": "Service unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Exactly one header. Scoped to authenticated organization, principal and operation. Reusing a key with changed input returns 409; retired receipts return 410. Retry unknown outcomes with the same key and input.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9_-]+$"
        }
      }
    }
  }
}